+++ src/test/java/nl/lexemmens/podman/SaveMojoTest.java
@@ -20,6 +20,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Path;
@@ -42,7 +43,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .setUseMavenProjectVersion(false)
                 .setContainerfileDir(DEFAULT_CONTAINERFILE_DIR)
                 .build();
-        configureMojo(image, true, false, null);
+        configureMojo(image, true, false, null, true);
 
         saveMojo.execute();
 
@@ -55,7 +56,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .setUseMavenProjectVersion(false)
                 .setContainerfileDir(DEFAULT_CONTAINERFILE_DIR)
                 .build();
-        configureMojo(image, false, true, null);
+        configureMojo(image, false, true, null, true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target/podman-test");
@@ -71,7 +72,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .setUseMavenProjectVersion(false)
                 .setContainerfileDir(DEFAULT_CONTAINERFILE_DIR)
                 .build();
-        configureMojo(image,false, false, null);
+        configureMojo(image,false, false, null, true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target/podman-test");
@@ -83,12 +84,34 @@ public class SaveMojoTest extends AbstractMojoTest {
     }
 
     @Test
+    public void testSaveWithoutContainerfileDoesNotThrowExceptionWhenConfigured() {
+        String containerFileDir = "src/test/non-existing-directory";
+        Path currentDir = Paths.get(".");
+        Path targetLocation = currentDir.resolve(containerFileDir);
+        String targetLocationAsString = targetLocation.normalize().toFile().getAbsolutePath();
+
+        ImageConfiguration image = new TestImageConfigurationBuilder("sample")
+                .setContainerfileDir(targetLocationAsString)
+                .setTags(new String[]{"1.0.0"})
+                .setCreateLatestTag(false)
+                .build();
+        configureMojo(image, false, false, "registry.example.com", false);
+
+        when(mavenProject.getBuild()).thenReturn(build);
+        when(build.getDirectory()).thenReturn("target");
+
+        Assertions.assertDoesNotThrow(saveMojo::execute);
+        verify(log, Mockito.times(1)).warn(Mockito.eq("No Containerfile was found at " + targetLocationAsString + File.separator + "Containerfile, however this will be ignored due to current plugin configuration."));
+        verify(log, Mockito.times(1)).warn(Mockito.eq("Skipping save of container image with name sample. Configuration is not valid for this module!"));
+    }
+
+    @Test
     public void testSaveWithMavenProjectVersion() throws MojoExecutionException {
         ImageConfiguration image = new TestImageConfigurationBuilder("sample")
                 .setUseMavenProjectVersion(true)
                 .setContainerfileDir(DEFAULT_CONTAINERFILE_DIR)
                 .build();
-        configureMojo(image,false, false, "registry.example.com");
+        configureMojo(image,false, false, "registry.example.com", true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target/podman-test");
@@ -109,7 +132,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .setUseMavenProjectVersion(true)
                 .setContainerfileDir(DEFAULT_CONTAINERFILE_DIR)
                 .build();
-        configureMojo(image, false, false, null);
+        configureMojo(image, false, false, null, true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target/podman-test");
@@ -133,7 +156,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .setTags(new String[]{"1.0.0"})
                 .setCreateLatestTag(false)
                 .build();
-        configureMojo(image, false, false, "registry.example.com");
+        configureMojo(image, false, false, "registry.example.com", true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target");
@@ -162,7 +185,7 @@ public class SaveMojoTest extends AbstractMojoTest {
                 .addCustomImageNameForBuildStage("phase", "image-name-number-1")
                 .addCustomImageNameForBuildStage("phase2", "image-name-number-2")
                 .build();
-        configureMojo(image, false, false, "registry.example.com");
+        configureMojo(image, false, false, "registry.example.com", true);
 
         when(mavenProject.getBuild()).thenReturn(build);
         when(build.getDirectory()).thenReturn("target");
@@ -181,7 +204,7 @@ public class SaveMojoTest extends AbstractMojoTest {
         verify(log, times(1)).info(Mockito.eq("Container images exported successfully."));
     }
 
-    private void configureMojo(ImageConfiguration image, boolean skipAll, boolean skipSave, String pushRegistry) {
+    private void configureMojo(ImageConfiguration image, boolean skipAll, boolean skipSave, String pushRegistry, boolean failOnMissingContainerfile) {
         List<ImageConfiguration> images = List.of(image);
 
         saveMojo.podman = new TestPodmanConfigurationBuilder().setTlsVerify(TlsVerify.NOT_SPECIFIED).build();
@@ -190,5 +213,6 @@ public class SaveMojoTest extends AbstractMojoTest {
         saveMojo.skipSave = skipSave;
         saveMojo.pushRegistry = pushRegistry;
         saveMojo.images = images;
+        saveMojo.failOnMissingContainerfile = failOnMissingContainerfile;
     }
 }